REFACTOR: split the code from contacts.js. by files
