+++ controllers/contacts/updateStatusContact.js
@@ -1,0 +1,13 @@
+const { Contact } = require("../../models");
+const { HttpError } = require("../../helpers");
+
+const updateStatusContact = async (req, res) => {
+  const { id } = req.params;
+  const data = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  if (!data) {
+    throw HttpError(404, "Not Found");
+  }
+  return res.status(200).json(data);
+};
+
+module.exports = updateStatusContact
